+++ NEWS.md
@@ -1,5 +1,7 @@
 # cli (development version)
 
+* `cli_progress_bar()` has a new argument `quiet`. If TRUE, then progress bar will be suppressed.
+
 * `keypress()` improvements:
   - `timeout` argument to wait at most a given number of seconds for a
     key press.